requirements.txt
mem0ai>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
opensearch-py>=2.0.0
boto3>=1.20.0
requests>=2.28.0
